Refactor/extract layout
